# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "目前仅仅是一个demo",
        "title": "Rest API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/AddUser": {
            "post": {
                "description": "新增用户信息 请求主体: base64(str=aaaa) 成功输出null",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "新增用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "str",
                        "name": "str",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/DeleteUser": {
            "post": {
                "description": "删除用户信息 请求主体: base64(str=aaaa) 成功输出null",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "删除用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "str",
                        "name": "str",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetAllUserInfo": {
            "post": {
                "description": "前台条件获取用户信息 请求主体: base64(ParameterStr=aaaa\u0026PageSize=10\u0026CurrentPage=1) 成功输出[]User",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "前台条件获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ParameterStr",
                        "name": "ParameterStr",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PageSize",
                        "name": "PageSize",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "CurrentPage",
                        "name": "CurrentPage",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]User},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/GetAllUserViewInfo": {
            "post": {
                "description": "前台条件获取用户信息 请求主体: base64(ParameterStr=aaaa\u0026PageSize=10\u0026CurrentPage=1) 成功输出[]User",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "前台条件获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ParameterStr",
                        "name": "ParameterStr",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PageSize",
                        "name": "PageSize",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "CurrentPage",
                        "name": "CurrentPage",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]UserView},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetUserByID": {
            "post": {
                "description": "根据用户ID获取用户信息 请求主体: base64(ID=aaaa) 成功输出User",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据用户ID获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "ID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{User},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetUserMenu": {
            "post": {
                "description": "根据用户ID获取用户菜单信息 请求主体:  base64(userID=aaaa)  成功输出[]MenuTree",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据用户ID获取用户菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "userID",
                        "name": "userID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]MenuTree},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/SetUserDept": {
            "post": {
                "description": "设置用户部门 请求主体: base64(ID=aaaa) 成功输出User",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "设置用户部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "部门ID，拼着接字符串",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/SetUserRole": {
            "post": {
                "description": "设置设置用户权限用户部门 请求主体: base64(ID=aaaa) 成功输出User",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "设置用户权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "权限ID，拼着接字符串",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/UpdateUser": {
            "post": {
                "description": "修改用户信息 请求主体: base64(str=aaaa) 成功输出null",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "修改用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "str",
                        "name": "str",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}